move train code to train.py and change bash script accordingly
